+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The changelog is applicable from version `4.6.0` onwards.
 * RAD-244: Move uberjar support into separate plugin `com.brightsparklabs.gradle.baseline-uberjar` (**BREAKING CHANGE**).
 * RAD-244: Convert README from Markdown to Asciidoc.
 * RAD-245: Replace deprecated 'Johnrengelman Shadow' plugin with the latest 'Gradleup Shadow' plugin.
+* RAD-246: Bumped Java version from 21 to 25 and Gradle to 9.4.0. Upgraded any broken dependencies due to this change. (**BREAKING CHANGE**).
 
 ### Deprecated