+++ CHANGELOG.md
@@ -8,13 +8,49 @@ The changelog is applicable from version `4.6.0` onwards.
 
 ---
 
+## [Unreleased] - YYYY-MM-DD
+
+[Unreleased]: https://github.com/brightsparklabs/gradle-baseline/compare/x.y.z...HEAD
+
+### Added
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
+
+---
+
+## [7.0.0] - 2026-03-16
+
+[Unreleased]: https://github.com/brightsparklabs/gradle-baseline/compare/6.0.0...7.0.0
+
+**NOTE: CONTAINS BREAKING CHANGES.**
+
+### Changed
+
+* RAD-244: Move uberjar support into separate plugin `com.brightsparklabs.gradle.baseline-uberjar` (**BREAKING CHANGE**).
+* RAD-244: Convert README from Markdown to Asciidoc.
+* RAD-245: Replace deprecated 'Johnrengelman Shadow' plugin with the latest 'Gradleup Shadow' plugin.
+* RAD-246: Bumped Java version from 21 to 25 and Gradle to 9.4.0.
+  Upgraded any broken dependencies due to this change (**BREAKING CHANGE**).
+
+---
+
 ## [6.0.0] - 2025-10-21
 
 [6.0.0]: https://github.com/brightsparklabs/gradle-baseline/compare/5.1.0...6.0.0
 
+**NOTE: CONTAINS BREAKING CHANGES.**
+
 ### Changed
 
-* RAD-237: Bumped Java version from 17 to 21.
+* RAD-237: Bumped Java version from 17 to 21 (**BREAKING CHANGE**).
 
 ## [5.1.0] - 2025-10-07
 
@@ -28,9 +64,11 @@ The changelog is applicable from version `4.6.0` onwards.
 
 [5.0.0]: https://github.com/brightsparklabs/gradle-baseline/compare/4.10.0...5.0.0
 
+**NOTE: CONTAINS BREAKING CHANGES.**
+
 ### Changed
 
-* RAD-237: Bumped Gradle major version from 8 to 9.
+* RAD-237: Bumped Gradle major version from 8 to 9 (**BREAKING CHANGE**).
 
 ## [4.10.0] - 2025-09-05